Fixes and optimizations in dataset generation
Sometimes, the dataset generation execution failed due to joblib
(when all the strategies were set, the fail was caused in the first
generator, which in the executions was the random generator), and
the reason was the used memory. Got the hint from
https://lightrun.com/answers/joblib-joblib-loky-terminatedworkererror

BOW intersection metric was not being correctly calculated. Now it
is using the Jaccard metric.